+++ test/arithmetic-expression/feel-arithmetic-expression.build.spec.js
@@ -7,7 +7,7 @@
 var chalk = require('chalk');
 var chai = require('chai');
 var expect = chai.expect;
-var FEEL = require('../dist/feel');
+var FEEL = require('../../dist/feel');
 
 describe(chalk.blue('Arithmetic expression ast parsing test'), function() {
     debugger;
