+++ test/arithmetic-expression/feel-arithmetic-expression.parse.spec.js
@@ -7,25 +7,12 @@
 var chalk = require('chalk');
 var chai = require('chai');
 var expect = chai.expect;
-var FEEL = require('../dist/feel');
+var FEEL = require('../../dist/feel');
 
-describe(chalk.blue('IF expression grammar test'), function () {
+describe(chalk.blue('Arithmetic expression grammar test'), function() {
 
-    it('Successfully creates ast from simple if expression', function (done) {
-        var text = 'if applicant.maritalStatus in ("M", "S") then "valid" else "not valid"';
-
-        try {
-            var parsedGrammar = FEEL.parse(text);
-            expect(parsedGrammar).not.to.be.undefined;
-        } catch (e) {
-            expect(parsedGrammar).not.to.be.undefined;
-            expect(e).to.be.undefined;
-        }
-        done();
-    });
-
-    it('Successfully creates ast from if expression', function (done) {
-        var text = 'if Pre-Bureau Risk Category = "DECLINE" or Installment Affordable = false or Age < 18 or Monthly Income < 100 then "INELIGIBLE" else "ELIGIBLE"';
+    it('Successfully creates ast from simple arithmetic expression', function(done) {
+        var text = 'a + b - c';
 
         try {
             var parsedGrammar = FEEL.parse(text);
@@ -37,9 +24,35 @@ describe(chalk.blue('IF expression grammar test'), function () {
         done();
     });
 
-    it('Successfully creates ast from if expression', function (done) {
-        var text = 'if "Pre-Bureau Risk Category" = "DECLINE" or "Installment Affordable" = false or Age < 18 or "Monthly Income" < 100 then "INELIGIBLE" else "ELIGIBLE"';
-        
+    it('Successfully creates ast from arithmetic expression', function(done) {
+        var text = '((a + b)/c - (d + e*2))**f';
+
+        try {
+            var parsedGrammar = FEEL.parse(text);
+            expect(parsedGrammar).not.to.be.undefined;
+        } catch (e) {
+            expect(parsedGrammar).not.to.be.undefined;
+            expect(e).to.be.undefined;
+        }
+        done();
+    });
+
+    it('Successfully creates ast from arithmetic expression', function(done) {
+        var text = '1-(1+rate/12)**-term';
+
+        try {
+            var parsedGrammar = FEEL.parse(text);
+            expect(parsedGrammar).not.to.be.undefined;
+        } catch (e) {
+            expect(parsedGrammar).not.to.be.undefined;
+            expect(e).to.be.undefined;
+        }
+        done();
+    });
+
+    it('Successfully creates ast from arithmetic expression', function(done) {
+        var text = '(a + b)**-c';
+
         try {
             var parsedGrammar = FEEL.parse(text);
             expect(parsedGrammar).not.to.be.undefined;
