+++ test/comparision-expression/feel-comparision-expression.build.spec.js
@@ -7,7 +7,7 @@
 var chalk = require('chalk');
 var chai = require('chai');
 var expect = chai.expect;
-var FEEL = require('../dist/feel');
+var FEEL = require('../../dist/feel');
 
 describe(chalk.blue('Comparision expression ast parsing test'), function() {
