+++ test/disjunction-conjunction-expression/feel-disjunction-conjunction.build.spec.js
@@ -7,7 +7,7 @@
 var chalk = require('chalk');
 var chai = require('chai');
 var expect = chai.expect;
-var FEEL = require('../dist/feel');
+var FEEL = require('../../dist/feel');
 
 describe(chalk.blue('Disjunction-Conjunction ast parsing test'), function() {
