+++ utils/helper/fn-negation.js
